Machine Learning/Fellow-Projects/Drought_and_Risk/requirements.txt
# Core ML and Scientific Computing
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
scikit-learn>=1.0.0

# Deep Learning
torch>=1.9.0
torchvision>=0.10.0

# Hydrology and Geospatial
netCDF4>=1.5.7
xarray>=0.19.0
pygeohydro>=0.13.0
pynhd>=0.13.0
geopandas>=0.10.0

# Data Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0

# Jupyter and Analysis
jupyter>=1.0.0
ipykernel>=6.0.0
notebook>=6.4.0

# Optimization and Hyperparameter Tuning
optuna>=2.10.0

# Configuration and Utilities
pyyaml>=5.4.0
tqdm>=4.62.0
requests>=2.26.0

# Testing
pytest>=6.2.0
pytest-cov>=2.12.0

# Code Quality
black>=21.0.0
flake8>=3.9.0
isort>=5.9.0

# Documentation
sphinx>=4.0.0
sphinx-rtd-theme>=0.5.0

# Additional utilities
python-dotenv>=0.19.0
click>=8.0.0
